+++ .ci/scripts/create_all_plugins_app.sh
@@ -3,5 +3,5 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-dart ./script/tool/bin/flutter_plugin_tools.dart create-all-packages-app \
+dart pub global run flutter_plugin_tools create-all-packages-app \
     --output-dir=. --exclude script/configs/exclude_all_packages_app.yaml