+++ packages/camera/camera_android_camerax/test/camera_selector_test.dart
@@ -60,10 +60,10 @@ void main() {
       );
       CameraSelector(
           instanceManager: instanceManager,
-          lensFacing: CameraSelector.LENS_FACING_BACK);
+          lensFacing: CameraSelector.lensFacingBack);
 
       verify(
-          mockApi.create(argThat(isA<int>()), CameraSelector.LENS_FACING_BACK));
+          mockApi.create(argThat(isA<int>()), CameraSelector.lensFacingBack));
     });
 
     test('filterTest', () async {
@@ -108,14 +108,14 @@ void main() {
         instanceManager: instanceManager,
       );
 
-      flutterApi.create(0, CameraSelector.LENS_FACING_BACK);
+      flutterApi.create(0, CameraSelector.lensFacingBack);
 
       expect(instanceManager.getInstanceWithWeakReference(0),
           isA<CameraSelector>());
       expect(
           (instanceManager.getInstanceWithWeakReference(0)! as CameraSelector)
               .lensFacing,
-          equals(CameraSelector.LENS_FACING_BACK));
+          equals(CameraSelector.lensFacingBack));
     });
   });
 }