+++ packages/file_selector/file_selector_linux/test/file_selector_linux_test.dart
@@ -16,10 +16,15 @@ void main() {
   setUp(() {
     plugin = FileSelectorLinux();
     log = <MethodCall>[];
-    plugin.channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      log.add(methodCall);
-      return null;
-    });
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      plugin.channel,
+      (MethodCall methodCall) async {
+        log.add(methodCall);
+        return null;
+      },
+    );
   });
 
   test('registers instance', () {
@@ -424,3 +429,9 @@ void expectMethodCall(
 }) {
   expect(log, <Matcher>[isMethodCall(methodName, arguments: arguments)]);
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;