+++ packages/file_selector/file_selector_platform_interface/test/method_channel_file_selector_test.dart
@@ -16,10 +16,15 @@ void main() {
     final List<MethodCall> log = <MethodCall>[];
 
     setUp(() {
-      plugin.channel.setMockMethodCallHandler((MethodCall methodCall) async {
-        log.add(methodCall);
-        return null;
-      });
+      _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+          .defaultBinaryMessenger
+          .setMockMethodCallHandler(
+        plugin.channel,
+        (MethodCall methodCall) async {
+          log.add(methodCall);
+          return null;
+        },
+      );
 
       log.clear();
     });
@@ -274,3 +279,9 @@ void expectMethodCall(
 }) {
   expect(log, <Matcher>[isMethodCall(methodName, arguments: arguments)]);
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;