+++ packages/google_sign_in/google_sign_in_platform_interface/test/method_channel_google_sign_in_test.dart
@@ -50,7 +50,9 @@ void main() {
 
     setUp(() {
       responses = Map<String, dynamic>.from(kDefaultResponses);
-      channel.setMockMethodCallHandler((MethodCall methodCall) {
+      _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+          .defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (MethodCall methodCall) {
         log.add(methodCall);
         final dynamic response = responses[methodCall.method];
         if (response != null && response is Exception) {
@@ -160,3 +162,9 @@ void main() {
     });
   });
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;