+++ packages/image_picker/image_picker_platform_interface/test/new_method_channel_image_picker_test.dart
@@ -18,7 +18,10 @@ void main() {
 
     setUp(() {
       returnValue = '';
-      picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+          .defaultBinaryMessenger
+          .setMockMethodCallHandler(picker.channel,
+              (MethodCall methodCall) async {
         log.add(methodCall);
         return returnValue;
       });
@@ -185,8 +188,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.pickImage(source: ImageSource.gallery), isNull);
         expect(await picker.pickImage(source: ImageSource.camera), isNull);
@@ -352,8 +357,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.pickMultiImage(), isNull);
         expect(await picker.pickMultiImage(), isNull);
@@ -424,8 +431,10 @@ void main() {
       });
 
       test('handles a null video path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.pickVideo(source: ImageSource.gallery), isNull);
         expect(await picker.pickVideo(source: ImageSource.camera), isNull);
@@ -467,7 +476,10 @@ void main() {
 
     group('#retrieveLostData', () {
       test('retrieveLostData get success response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'image',
             'path': '/example/path',
@@ -480,7 +492,10 @@ void main() {
       });
 
       test('retrieveLostData get error response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'video',
             'errorCode': 'test_error_code',
@@ -495,14 +510,20 @@ void main() {
       });
 
       test('retrieveLostData get null response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return null;
         });
         expect((await picker.retrieveLostData()).isEmpty, true);
       });
 
       test('retrieveLostData get both path and error should throw', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'video',
             'errorCode': 'test_error_code',
@@ -672,8 +693,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.getImage(source: ImageSource.gallery), isNull);
         expect(await picker.getImage(source: ImageSource.camera), isNull);
@@ -839,8 +862,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.getMultiImage(), isNull);
         expect(await picker.getMultiImage(), isNull);
@@ -911,8 +936,10 @@ void main() {
       });
 
       test('handles a null video path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.getVideo(source: ImageSource.gallery), isNull);
         expect(await picker.getVideo(source: ImageSource.camera), isNull);
@@ -954,7 +981,10 @@ void main() {
 
     group('#getLostData', () {
       test('getLostData get success response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'image',
             'path': '/example/path',
@@ -967,7 +997,10 @@ void main() {
       });
 
       test('getLostData should successfully retrieve multiple files', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, dynamic>{
             'type': 'image',
             'path': '/example/path1',
@@ -983,7 +1016,10 @@ void main() {
       });
 
       test('getLostData get error response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'video',
             'errorCode': 'test_error_code',
@@ -998,14 +1034,20 @@ void main() {
       });
 
       test('getLostData get null response', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return null;
         });
         expect((await picker.getLostData()).isEmpty, true);
       });
 
       test('getLostData get both path and error should throw', () async {
-        picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(picker.channel,
+                (MethodCall methodCall) async {
           return <String, String>{
             'type': 'video',
             'errorCode': 'test_error_code',
@@ -1201,8 +1243,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.getImageFromSource(source: ImageSource.gallery),
             isNull);
@@ -1431,8 +1475,10 @@ void main() {
       });
 
       test('handles a null image path response gracefully', () async {
-        picker.channel
-            .setMockMethodCallHandler((MethodCall methodCall) => null);
+        _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+            .defaultBinaryMessenger
+            .setMockMethodCallHandler(
+                picker.channel, (MethodCall methodCall) => null);
 
         expect(await picker.getMultiImage(), isNull);
         expect(await picker.getMultiImage(), isNull);
@@ -1478,3 +1524,9 @@ void main() {
     });
   });
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;