+++ packages/local_auth/local_auth_ios/test/local_auth_test.dart
@@ -18,7 +18,9 @@ void main() {
     late LocalAuthIOS localAuthentication;
 
     setUp(() {
-      channel.setMockMethodCallHandler((MethodCall methodCall) {
+      _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+          .defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (MethodCall methodCall) {
         log.add(methodCall);
         switch (methodCall.method) {
           case 'getEnrolledBiometrics':
@@ -181,3 +183,9 @@ void main() {
     });
   });
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;