+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,4 +1,6 @@
-## NEXT
+## 0.8.6+2
+
+* Updates `NSPhotoLibraryUsageDescription` description in README.
 
 * Updates minimum Flutter version to 3.0.